refac: only register fonts when module is loaded
Instead of registering the font on every render,
it now only registers the fonts when the module is loaded.
