+++ src/HTMLPanel.tsx
@@ -17,7 +17,7 @@ import { triggerPanelwillunmount } from 'utils/events/panelwillunmount';
 import { addHtml } from 'utils/addHtml';
 import { CustomScrollbar, ScrollContainer } from '@grafana/ui';
 import semver from 'semver';
-import { css } from '@emotion/css';
+import './registerFonts';
 
 interface Props extends PanelProps<OptionsInterface> {}
 interface PanelState {
@@ -298,8 +298,6 @@ export class HTMLPanel extends PureComponent<Props, PanelState> {
       options: { useGrafanaScrollbar, overflow },
     } = this.props;
 
-    registerFonts();
-
     return (
       <>
         <div style={{ position: 'absolute', width: `${width}px`, height: `${height}px` }}>
@@ -322,45 +320,4 @@ export class HTMLPanel extends PureComponent<Props, PanelState> {
       </>
     );
   }
-}
-
-function registerFonts() {
-  css([
-    {
-      '@font-face': {
-        fontFamily: 'Open Sans',
-        src: `url('${__webpack_public_path__}static/fonts/OpenSans-Light.woff2') format('woff2')`,
-        fontWeight: '300',
-        fontStyle: 'normal',
-        fontDisplay: 'swap',
-      },
-    },
-    {
-      '@font-face': {
-        fontFamily: 'Open Sans',
-        src: `url('${__webpack_public_path__}static/fonts/OpenSans-Regular.woff2') format('woff2')`,
-        fontWeight: 'normal',
-        fontStyle: 'normal',
-        fontDisplay: 'swap',
-      },
-    },
-    {
-      '@font-face': {
-        fontFamily: 'Open Sans',
-        src: `url('${__webpack_public_path__}static/fonts/OpenSans-SemiBold.woff2') format('woff2')`,
-        fontWeight: '600',
-        fontStyle: 'normal',
-        fontDisplay: 'swap',
-      },
-    },
-    {
-      '@font-face': {
-        fontFamily: 'Open Sans',
-        src: `url('${__webpack_public_path__}static/fonts/OpenSans-Bold.woff2') format('woff2')`,
-        fontWeight: 'bold',
-        fontStyle: 'normal',
-        fontDisplay: 'swap',
-      },
-    },
-  ]);
 }
